+++ shiny/ui/_chat.py
@@ -19,6 +19,7 @@ from weakref import WeakValueDictionary
 from htmltools import HTML, Tag, TagAttrValue, css
 
 from .. import _utils, reactive
+from .._deprecated import warn_deprecated
 from .._docstring import add_example
 from .._namespaces import ResolvedId, resolve_id
 from ..session import require_active_session, session_context
@@ -933,26 +934,43 @@ class Chat:
         id = self.user_input_id
         return cast(str, self._session.input[id]())
 
-    def set_user_message(self, value: str):
+    def update_user_input(
+        self, *, value: str | None = None, placeholder: str | None = None
+    ):
         """
-        Set the user's message.
+        Update the user input.
 
         Parameters
         ----------
         value
             The value to set the user input to.
+        placeholder
+            The placeholder text for the user input.
         """
+
+        obj = _utils.drop_none({"value": value, "placeholder": placeholder})
 
         _utils.run_coro_sync(
             self._session.send_custom_message(
                 "shinyChatMessage",
                 {
                     "id": self.id,
-                    "handler": "shiny-chat-set-user-input",
-                    "obj": value,
+                    "handler": "shiny-chat-update-user-input",
+                    "obj": obj,
                 },
             )
         )
+
+    def set_user_message(self, value: str):
+        """
+        Deprecated. Use `update_user_input(value=value)` instead.
+        """
+
+        warn_deprecated(
+            "set_user_message() is deprecated. Use update_user_input(value=value) instead."
+        )
+
+        self.update_user_input(value=value)
 
     async def clear_messages(self):
         """